+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.165 ] - 2025/02/17
+
+  * Updated `SpotifyClient.PlayerMediaPlayContext` and `PlayerMediaPlayTracks` to utilize Spotify Web Player authorization token when starting play for Sonos and restricted devices.  This will cause play to start under Spotify Connect control, rather than using the Sonos Local Queue.  Refer to the [Spotify Web Player Authentication Setup](https://github.com/thlucas1/homeassistantcomponent_spotifyplus/wiki/Device-Configuration-Options#spotify-web-player-authentication-setup) wiki page for more information on how to configure it.
+  
 ###### [ 1.0.164 ] - 2025/02/05
 
   * Added `SpotifyClient.Dispose` method to properly dispose of resources and unwire events when the instance is destroyed.friendly.